[fix] divide float으로 형변환

+++ src/Class_1005/Calculator.java
@@ -13,7 +13,7 @@ public class Calculator {
         return a*b;
     }
 
-    public int divide(int a, int b){
-        return a/b;
+    public float divide(int a, int b){
+        return (float) a / b;
     }
 }
